+++ R/amplicon.R
@@ -46,7 +46,66 @@ seqtab_stats <- function(seqtab) {
 
 
 
+#' Stats of multiple seqtab objects
+#'
+#' @param ... One or more seqtab objects
+#'
+#' @returns A tibble with seqtab stats
+#' @export
+
+seqtab_stats_bind <- function(...) {
+
+  seqtab_names <- as.character(substitute(list(...)))[-1]
+  seqtab_list <- mget(seqtab_names, envir = parent.frame())
+
+  res <- dplyr::bind_rows(
+    lapply(
+      seqtab_list,
+      seqtab_stats
+    ),
+    .id = "STEP"
+  ) |>
+    tidyr::pivot_longer(
+      cols = c("FREQ", "ABUNDANCE"),
+      names_to = "METRIC",
+      values_to = "VALUE"
+    )
+
+  return(res)
+}
 
 
 
+#' Plot multiple seqtab objects
+#'
+#' @param ... One or more seqtab objects
+#'
+#' @returns A ggplot object
+#' @export
 
+seqtab_stats_plot = function(...) {
+
+  # note, the top half of this (before the ggplot call) is identical to seqtab_stats_bind, but I can't use that function here because the mget call doesn't work
+
+  seqtab_names <- as.character(substitute(list(...)))[-1]
+  seqtab_list <- mget(seqtab_names, envir = parent.frame())
+
+  dplyr::bind_rows(
+    lapply(
+      seqtab_list,
+      seqtab_stats
+    ),
+    .id = "STEP"
+  ) |>
+    tidyr::pivot_longer(
+      cols = c("FREQ", "ABUNDANCE"),
+      names_to = "METRIC",
+      values_to = "VALUE"
+    ) |> #
+    ggplot2::ggplot(ggplot2::aes(x = LENGTH, y = VALUE, fill = STEP)) +
+      ggplot2::geom_col(position = "dodge") +
+      ggplot2::scale_fill_manual(values = palette_jak(n = 2, p = "bay")) +
+      ggplot2::facet_wrap(~METRIC, scales = "free_y", ncol = 1) +
+      ggplot2::scale_y_continuous(labels=function(x) format(x, big.mark = ",", decimal.mark = ".", scientific = FALSE)) +
+      ggplot2::labs(x = "ASV Length", y = "ASV Count", title = "ASV frequency and abundance by length")
+}